준수님_0603PR입니다!

+++ src/main/java/com/team/berp/order/dto/OrderItemRequest.java
@@ -7,5 +7,5 @@ public class OrderItemRequest {
     private Long itemId;
     private String unit;
     private Long unitPrice;
-    private Long unitQty;
+    private Integer unitQty;
 }
